+++ src/wg-tui.c
@@ -6,10 +6,11 @@
  */
 
 #include <stddef.h>
-#include <stdio.h>
 #include <string.h>
+#include <getopt.h>
 
 #include <configure.h>
+#include <wg-tui.h>
 
 #include "subcommands.h"
 
@@ -66,4 +67,81 @@ int main(int argc, const char *argv[])
 	fprintf(stderr, "Invalid subcommand: `%s'\n", argv[1]);
 	show_usage(stderr);
 	return 1;
+}
+
+int print_help(FILE *file)
+{
+	fprintf(file, "Usage: %s <cmd> [<args>]\n", PROG_NAME);
+	printf("args:");
+	printf("    -h, --help          Display help information and exit");
+	printf("    -u, --build-info    Display build information end exit");
+	return 0;
+}
+
+int print_build_info(FILE *file)
+{
+    char *use_bashcompletion = "NO";
+    char *use_wgquick = "NO";
+    char *use_systemdunits = "NO";
+
+#ifdef WG_TUI_WITH_BASHCOMPLETION
+    use_bashcompletion = "YES";
+#endif
+
+#ifdef WG_TUI_WITH_WGQUICK
+    use_wgquick = "YES";
+#endif
+
+#ifdef WG_TUI_WITH_SYSTEMDUNITS
+    use_systemdunits = "YES";
+#endif
+
+	fprintf(file, "Version: %s\n", WG_TUI_VERSION);
+	fprintf(file, "  Major: %s\n", WG_TUI_VERSION_MAJOR);
+	fprintf(file, "  Minor: %s\n", WG_TUI_VERSION_MINOR);
+	fprintf(file, "  Patch: %s\n", WG_TUI_VERSION_PATCH);
+	fprintf(file, "\n");
+	fprintf(file, "Description: %s\n", WG_TUI_DESCRIPTION);
+	fprintf(file, "Homepage url: %s\n", WG_TUI_HOMEPAGE_URL);
+	fprintf(file, "\n");
+	fprintf(file, "Project generator: %s\n", WG_TUI_GENERATOR);
+	fprintf(file, "Project build type: %s\n", WG_TUI_BUILD_TYPE);
+	fprintf(file, "\n");
+	fprintf(file, "C compiler variables:\n");
+	fprintf(file, "  Extensions: %s\n", WG_TUI_C_EXTENSIONS);
+	fprintf(file, "  Standard: %s\n", WG_TUI_C_STANDARD);
+	fprintf(file, "  Compiler: %s\n", WG_TUI_C_COMPILER);
+	fprintf(file, "  Compiler id: %s\n", WG_TUI_C_COMPILER_ID);
+	fprintf(file, "  Compiler predefines command: %s\n", WG_TUI_C_COMPILER_PREDEFINES_COMMAND);
+	fprintf(file, "  Compiler target: %s\n", WG_TUI_C_COMPILER_TARGET);
+    fprintf(file, "  Compiler version: %s\n", WG_TUI_C_COMPILER_VERSION);
+    fprintf(file, "  Flags: %s\n", WG_TUI_C_FLAGS);
+    fprintf(file, "  Simulate id: %s\n", WG_TUI_C_SIMULATE_ID);
+    fprintf(file, "  Standard include derictories: %s\n", WG_TUI_C_STANDARD_INCLUDE_DIRECTORIES);
+    fprintf(file, "  Standard libraries: %s\n", WG_TUI_C_STANDARD_LIBRARIES);
+    fprintf(file, "  Standard required: %s\n", WG_TUI_C_STANDARD_REQUIRED);
+    fprintf(file, "  Compiler ABI: %s\n", WG_TUI_C_COMPILER_ABI);
+    fprintf(file, "  Compiler architecture id: %s\n", WG_TUI_C_COMPILER_ARCHITECTURE_ID);
+    fprintf(file, "  Compiler compiler version internal: %s\n", WG_TUI_C_COMPILER_VERSION_INTERNAL);
+    fprintf(file, "  Linker preference: %s\n", WG_TUI_C_LINKER_PREFERENCE);
+    fprintf(file, "  Linker preference propagates: %s\n", WG_TUI_C_LINKER_PREFERENCE_PROPAGATES);
+    fprintf(file, "  Platform id: %s\n", WG_TUI_C_PLATFORM_ID);
+    fprintf(file, "\n");
+    fprintf(file, "CMake version: %s\n", WG_TUI_CMAKE_VERSION);
+    fprintf(file, "        Major: %s\n", WG_TUI_CMAKE_VERSION_MAJOR);
+    fprintf(file, "        Minor: %s\n", WG_TUI_CMAKE_VERSION_MINOR);
+    fprintf(file, "        Patch: %s\n", WG_TUI_CMAKE_VERSION_PATCH);
+    fprintf(file, "\n");
+    fprintf(file, "Application options:\n");
+    fprintf(file, "  Prefix: %s\n", PREFIX);
+    fprintf(file, "  Binary directory: %s\n", BINDIR);
+    fprintf(file, "  Library directory: %s\n", LIBDIR);
+    fprintf(file, "  Man pages directory: %s\n", MANDIR);
+    fprintf(file, "  Bash completion scripts: %s\n", BASHCOMPDIR);
+    fprintf(file, "  Run state directory: %s\n", RUNSTATEDIR);
+    fprintf(file, "  Use bash completeon scripts: %s\n", use_bashcompletion);
+    fprintf(file, "  Use wg quick: %s\n", use_wgquick);
+    fprintf(file, "  Use systemd units: %s\n", use_systemdunits);
+
+	return 0;
 }